refactor(proxmox_firewall): rename modules to clarify usage
