+++ changelogs/fragments/450-rename-modules-proxmox-firewall-options.yml
@@ -1,0 +1,5 @@
+---
+minor_changes:
+  - proxmox_node_firewall - rename to proxmox_node_firewall_options (https://github.com/ansible-collections/community.proxmox/issues/450).
+  - proxmox_node_firewall_info - rename to proxmox_node_firewall_options_info (https://github.com/ansible-collections/community.proxmox/issues/450).
+  - proxmox_cluster_firewall - rename to proxmox_cluster_firewall_options (https://github.com/ansible-collections/community.proxmox/issues/450).
