+++ meta/runtime.yml
@@ -26,7 +26,7 @@ action_groups:
     - proxmox_ceph_osd
     - proxmox_ceph_pool
     - proxmox_cluster
-    - proxmox_cluster_firewall
+    - proxmox_cluster_firewall_options
     - proxmox_cluster_firewall_security_group
     - proxmox_cluster_ha_resources
     - proxmox_cluster_ha_rules
@@ -45,8 +45,8 @@ action_groups:
     - proxmox_ipam_info
     - proxmox_kvm
     - proxmox_nic
-    - proxmox_node_firewall_info
-    - proxmox_node_firewall
+    - proxmox_node_firewall_options_info
+    - proxmox_node_firewall_options
     - proxmox_node
     - proxmox_node_info
     - proxmox_node_network
